:star2: feat(global): wip converting to react-router 4
beginning the process of converting to RR4. For the most part, the majority of the routing is
complete. Data fetching isn't fully there in the dashboard. Theres a :bear: of a problem with the
dashboard routes not rendering as children. the sidebar goes away and the whole page re-renders on
route change.